Reworked exposoin effect
Fixed TurnipHelicopter not moving, seems to have been related to the shadow animation being in the turnipHelictoper controller without any transitions. I simply removed it.

Added a few more enemies, and noticed the waves spawning is odd, the later waves comes very quickly, 1 min from 1 to 2, then just 5 or 10 seconds for 3, 4, and 5.